Move things to separate files; faster metrics; new metrics
