+++ seismiqb/src/__init__.py
@@ -1,6 +1,8 @@
 """Init file"""
 from .utils import * # pylint: disable=wildcard-import
 from .plot_utils import * # pylint: disable=wildcard-import
+from .labels_utils import * # pylint: disable=wildcard-import
+from .horizon import * # pylint: disable=wildcard-import
 from .crop_batch import SeismicCropBatch
 from .cubeset import SeismicCubeset
 from .geometry import SeismicGeometry
