+++ seismiqb/src/_const.py
@@ -1,0 +1,4 @@
+""" Container for constant values used in SeismiQB. """
+
+FILL_VALUE = -999 # placeholder value for horizon depth inside labels dictionary
+FILL_VALUE_MAP = -999999 # placeholder value for horizon depth inside depth map
